// packages/vgraph/src/components/legend/type.ts
export type CategoryLegendDataItem = {
  /**
   * The shape configurations of a legend item.
   * 图例中图形配置
   */
  marker?: {
    /**
     * The type of shape of a legend item.
     * 图例中图形的类型
     */
    type?:
      | "rect"
      | "circle"
      | "icon"
      | "image"
      | "line"
      | "quadratic"
      | "cubic";
    [key: string]: unknown;
  };
  /**
   * The label configurations of a legend item.
   * 图例中文本标签的配置
   */
  label?: string | TextConfigs;
  /**
   * The value of the legend item.
   * 图例对应值
   */
  value?: string | number;
};